+++ CHANGES.md
@@ -13,6 +13,7 @@ Changes in 6.9
 - Only list Teacher Course Periods & set UserCoursePeriod() in Side.php & TeacherPrograms.php
 - Add School Periods select input in SchoolPeriodsSelectInput.fnc.php, TakeAttendance.php & EnterEligibility.php
 - Add "Year and Semester marking periods are not graded." note in Grades/Configuration.php
+- Add "You already have entered eligibility this week for this course period." note in EnterEligibility.php
 
 Changes in 6.8.1
 ----------------